POC of ExtendedLeavesBlock with separate DISTANCE property.
This is just to show it can work (with minor changes to Terrestria) and allow further discussion.  At the moment I do not intend to merge this change because I am certain it will negatively affect mod compatibility.  Any mod which assumes it can fetch LeavesBlock.DISTANCE for any block that is an instanceof LeavesBlock will crash.  Likewise for assuming about blocks in the "minecraft:leaves" tag.